requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
gunicorn==20.1.0

